Introduced Board Interface
Board class implementation renamed. Classes arranged in proper packages
